refactor: movendo arquivos do pacote br.edu.ifg.numbers para br.edu.ifg.numbers.gpatri.
