+++ ms-usuarios/src/main/java/br/edu/ifg/numbers/gpatri/msusuarios/MsUsuariosApplication.java
@@ -1,4 +1,4 @@
-package br.edu.ifg.numbers.msusuarios;
+package br.edu.ifg.numbers.gpatri.msusuarios;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
