+++ ms-usuarios/src/main/java/br/edu/ifg/numbers/gpatri/msusuarios/dto/CargoRequestDTO.java
@@ -1,4 +1,4 @@
-package br.edu.ifg.numbers.msusuarios.dto;
+package br.edu.ifg.numbers.gpatri.msusuarios.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
