+++ ms-usuarios/src/main/java/br/edu/ifg/numbers/gpatri/msusuarios/exception/ConflictException.java
@@ -1,4 +1,4 @@
-package br.edu.ifg.numbers.msusuarios.exception;
+package br.edu.ifg.numbers.gpatri.msusuarios.exception;
 
 public class ConflictException extends RuntimeException{
