+++ ms-usuarios/src/main/java/br/edu/ifg/numbers/gpatri/msusuarios/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package br.edu.ifg.numbers.msusuarios.exception;
+package br.edu.ifg.numbers.gpatri.msusuarios.exception;
 
 public class ResourceNotFoundException extends RuntimeException{
