+++ ms-usuarios/src/main/java/br/edu/ifg/numbers/gpatri/msusuarios/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package br.edu.ifg.numbers.msusuarios.repository;
+package br.edu.ifg.numbers.gpatri.msusuarios.repository;
 
-import br.edu.ifg.numbers.msusuarios.domain.Usuario;
+import br.edu.ifg.numbers.gpatri.msusuarios.domain.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
