+++ ms-usuarios/src/test/java/br/edu/ifg/numbers/gpatri/msusuarios/MsUsuariosApplicationTests.java
@@ -1,4 +1,4 @@
-package br.edu.ifg.numbers.msusuarios;
+package br.edu.ifg.numbers.gpatri.msusuarios;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
